Move value and minimax into core package
