+++ catanatron_experimental/catanatron_experimental/cli/accumulators.py
@@ -25,8 +25,8 @@ from catanatron_experimental.machine_learning.utils import (
     populate_matrices,
     DISCOUNT_FACTOR,
 )
-from catanatron.gym.features import create_sample
-from catanatron.gym.envs.catanatron_gym import to_action_space, to_action_type_space
+from catanatron.features import create_sample
+from catanatron.gym.envs.catanatron_env import to_action_space, to_action_type_space
 from catanatron.gym.board_tensor_features import (
     create_board_tensor,
 )
