+++ catanatron_experimental/catanatron_experimental/machine_learning/players/online_mcts_dqn.py
@@ -25,7 +25,7 @@ from tensorflow.keras.optimizers import Adam
 from catanatron.game import Game
 from catanatron.models.player import Player
 from catanatron_experimental.machine_learning.players.playouts import run_playouts
-from catanatron.gym.features import (
+from catanatron.features import (
     create_sample_vector,
     get_feature_ordering,
 )
