+++ catanatron_experimental/catanatron_experimental/machine_learning/players/reinforcement.py
@@ -6,12 +6,12 @@ from tensorflow import keras
 
 from catanatron.models.player import Player
 from catanatron.models.enums import Action, ActionType
-from catanatron.gym.features import (
+from catanatron.features import (
     create_sample,
     create_sample_vector,
     get_feature_ordering,
 )
-from catanatron.gym.envs.catanatron_gym import ACTIONS_ARRAY, ACTION_SPACE_SIZE
+from catanatron.gym.envs.catanatron_env import ACTIONS_ARRAY, ACTION_SPACE_SIZE
 from catanatron.gym.board_tensor_features import (
     NUMERIC_FEATURES,
     create_board_tensor,
