+++ catanatron_experimental/catanatron_experimental/mcts_score_collector.py
@@ -15,9 +15,9 @@ from catanatron.state_functions import (
     player_num_dev_cards,
     player_num_resource_cards,
 )
-from catanatron_experimental.machine_learning.players.minimax import AlphaBetaPlayer
+from catanatron.players.minimax import AlphaBetaPlayer
 from catanatron_experimental.machine_learning.players.playouts import run_playouts
-from catanatron.gym.features import (
+from catanatron.features import (
     build_production_features,
     resource_hand_features,
 )
