+++ catanatron_experimental/catanatron_experimental/rayopt.py
@@ -1,7 +1,7 @@
 from ray import tune
 
 from catanatron.models.player import Color
-from catanatron_experimental.machine_learning.players.value import (
+from catanatron.players.value import (
     DEFAULT_WEIGHTS,
     ValueFunctionPlayer,
 )
