+++ catanatron_server/catanatron_server/api.py
@@ -8,8 +8,8 @@ from catanatron_server.models import upsert_game_state, get_game_state
 from catanatron.json import GameEncoder, action_from_json
 from catanatron.models.player import Color, RandomPlayer
 from catanatron.game import Game
-from catanatron_experimental.machine_learning.players.value import ValueFunctionPlayer
-from catanatron_experimental.machine_learning.players.minimax import AlphaBetaPlayer
+from catanatron.players.value import ValueFunctionPlayer
+from catanatron.players.minimax import AlphaBetaPlayer
 from catanatron_experimental.analysis.mcts_analysis import GameAnalyzer
 
 bp = Blueprint("api", __name__, url_prefix="/api")
